+++ SpringBootMeli/src/main/java/com/springboot/meli/SpringBootMeli/entity/Search/Value.java
@@ -3,7 +3,7 @@ package com.springboot.meli.SpringBootMeli.entity.Search;
 public class Value{
     public String name;
     public Struct struct;
-    public int source;
+    public long source;
     public Object id;
     public int results;
 }
